updated user score component
