+++ src/gql/Score/queries.js
@@ -1,8 +1,8 @@
 import {gql} from 'react-apollo'
-import {levels} from './score-system'
+import {levels} from 'lib/score-system'
 
 export const COMMUNITY_SCORE_COUNTS_QUERY = gql`
-  query {
+  query communityScoreCounts{
     ${levels.one.name}: _allScoresMeta(filter:{value: ${levels.one.value}}){
           count
     }
@@ -10,6 +10,30 @@ export const COMMUNITY_SCORE_COUNTS_QUERY = gql`
       count
     }
     ${levels.three.name}: _allScoresMeta(filter:{value: ${levels.three.value}}){
+      count
+    }
+  }
+`
+
+export const USER_SCORE_COUNTS_QUERY = gql`
+  query userScoreCounts($scorecardId: ID){
+    ${levels.one.name}: _allScoresMeta(filter:{
+      value: ${levels.one.value},
+      scorecard:{id:$scorecardId}
+
+    }){
+          count
+    }
+    ${levels.two.name}: _allScoresMeta(filter:{
+      value:${levels.two.value},
+      scorecard:{id:$scorecardId}
+    }){
+      count
+    }
+    ${levels.three.name}: _allScoresMeta(filter:{
+      value: ${levels.three.value},
+      scorecard:{id:$scorecardId}
+    }){
       count
     }
   }
